requirements.txt
fastapi[standard]
pywebview
pyinstaller